# openapi/openapi-kafka-writer.yaml
openapi: 3.0.3
info:
  title: Kafka Writer API
  version: 1.0.1
  description: >
    Kafka Writer is a lightweight REST service responsible for producing
    messages to Kafka topics.

    It exposes HTTP endpoints that allow:
      - writing messages to Kafka topics (auto-creating them if missing)
      - creating Kafka topics on demand

servers:
  - url: http://localhost:5000
    description: Kafka Writer Service

tags:
  - name: kafka
    description: Kafka write and topic management operations

paths:
  /write:
    post:
      tags: [kafka]
      summary: Write message to Kafka topic
      description: >
        Produces a single message to a Kafka topic.
        
        If the topic does not exist, it is created automatically.
        The request body may contain arbitrary JSON, which is sent
        as the Kafka message payload.
      parameters:
        - name: topic
          in: query
          required: true
          schema:
            type: string
          description: Target Kafka topic name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Arbitrary JSON payload to send to Kafka
              type: object
      responses:
        "200":
          description: Message successfully written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteResponse"
        "400":
          description: Missing `topic` query parameter or invalid JSON
        "500":
          description: Kafka produce error

  /topic:
    post:
      tags: [kafka]
      summary: Create Kafka topic
      description: >
        Creates a new Kafka topic with a single partition
        and replication factor equal to one.

        If the topic already exists, no error is returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTopicRequest"
      responses:
        "200":
          description: Topic created or already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTopicResponse"
        "400":
          description: Topic creation failed

components:
  schemas:
    WriteResponse:
      type: object
      properties:
        message:
          type: string
          example: ok

    CreateTopicRequest:
      type: object
      required:
        - topic
      properties:
        topic:
          type: string
          description: Kafka topic name to create
          example: test-topic

    CreateTopicResponse:
      type: object
      properties:
        message:
          type: string
          example: "Topic 'test-topic' created."